# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Mandara API
  version: 0.1.0
  description: |
    Mandara is a self-hosted light-novel reading server.

    This document is the single source of truth for the HTTP API contract:
    the backend implements it and the frontend generates its types from it
    (see `frontend/package.json`, script `api-types`).

    Model overview: `books` are pure metadata entries; `book_files` are the
    actual books (uploads or virtual plugin books). Several files can share
    one metadata entry. Chapters, reading sessions and shares attach to
    files. Every file is either `private` (owner + admins only) or `public`
    (all logged-in users).
servers:
  - url: http://127.0.0.1:8080
    description: Local development server (default)
...
paths:
...
  /api/books:
    get:
      tags: [books]
      summary: List visible books with their visible files
      operationId: listBooks
      security:
        - bearerAuth: []
      parameters:
        - name: q
          in: query
          description: Substring match on book title
          required: false
          schema:
            type: string
        - name: source
          in: query
          description: Filter by file source ('local' or a plugin id)
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Metadata entries with at least one visible file, plus metadata-only entries (no files yet) the caller owns
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BookListEntry"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [books]
      summary: |
        Add a book to the library — the single acquisition flow
        (获取书籍 → 添加元数据). The content is either an uploaded
        `file`, a plugin book (`plugin_source` + `plugin_book_id`
        without a `file`; the plugin supplies the book file via
        `get-book-file` when available, otherwise chapters materialize
        lazily), or none (`metadata_only`: a bare metadata entry — add
        the content later by attaching a file/plugin book to it). The
        metadata entry is created automatically (parsed from the file /
        the plugin's own entry), taken from a plugin catalog
        (`meta_plugin_source` + `meta_plugin_book_id`), attached to an
        existing entry (`book_id`), or described manually. The metadata
        source is independent of the content source: metadata may come
        from plugin A while the content comes from plugin B.
        `title`/`authors`/`description`/`cover_url`/`ext` override the
        produced metadata (never applied when attaching). When the
        plugin source declares >1 volumes, the acquisition auto-splits
        into one series + one book per 卷 (the response carries them all).
        The split is driven by the CONTENT source's own volume
        declaration; the content source must declare its volume structure
        (400 otherwise — an undeclared content source cannot be sliced
        per volume) and may declare FEWER volumes than the metadata
        source (metadata-only series entries, e.g. bangumi 系列, may list
        not-yet-published volumes; those become metadata-only books with
        no content). A content source declaring MORE volumes than the
        metadata series is refused (400).
      operationId: uploadBook
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: |
                    Epub or txt file — the content. Mutually exclusive
                    with `plugin_source` + `plugin_book_id` (plugin
                    content) and with `metadata_only`: exactly one content
                    source is used. When given together with
                    `plugin_source`, the file is the content and the
                    plugin only supplies the metadata.
                metadata_only:
                  type: boolean
                  default: false
                  description: |
                    Create a bare metadata entry without any content
                    (no file is created). `meta_plugin_source` +
                    `meta_plugin_book_id` name the metadata source
                    (catalog entry), or `title`/`authors`/… describe the
                    entry manually (a title is required). Content may be
                    attached later via `POST /api/books/{id}/files`.
                visibility:
                  type: string
                  enum: [private, public]
                  default: private
                  description: Choose whether this file is published
                label:
                  type: string
                  description: Short human label for this edition (e.g. "proofread")
                book_id:
                  type: string
                  description: |
                    Attach the content to this existing metadata entry
                    instead of creating one. The caller must be able to
                    see the book. Mutually exclusive with the plugin
                    metadata modes (a plugin book attached this way must
                    not already be materialized under another entry).
                plugin_source:
                  type: string
                  description: |
                    Together with `plugin_book_id`: the *content* comes
                    from this plugin source's book (materialized into the
                    library; whole-book files via `get-book-file`, else
                    lazy chapter mode). When no `file`,
                    `meta_plugin_source`, or `book_id` is given, the
                    plugin's own entry also provides the metadata.
                plugin_book_id:
                  type: string
                  description: Id of the book inside the plugin's catalog
                meta_plugin_source:
                  type: string
                  description: |
                    Together with `meta_plugin_book_id`: the metadata
                    comes from this plugin source's catalog — possibly a
                    different plugin than the content source. Overrides
                    (`title`/…) apply on top of its entry.
                meta_plugin_book_id:
                  type: string
                  description: Id of the metadata book inside its plugin's catalog
                title:
                  type: string
                  description: "Manual metadata: overrides the title (auto mode uses the parsed/identified one)"
                authors:
                  type: string
                  description: 'Manual metadata: JSON string array, e.g. `["A","B"]`'
                description:
                  type: string
                  description: "Manual metadata: overrides the description"
                cover_url:
                  type: string
                  description: "Manual metadata: cover image URL"
                ext:
                  type: string
                  description: |
                    Manual metadata: extended fields as a JSON object
                    (see `BookExt`), e.g.
                    `{"isbn":"9787536692930","publisher":"出版社"}`.
                    Overrides the parsed/identified value per-key.
      responses:
        "201":
          description: |
            The created series (when a multi-volume source book was
            split) and the created books, each with its first file
            (empty files array for a metadata-only addition).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AcquireResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/books/{id}:
    get:
      tags: [books]
      summary: Book metadata plus files visible to the caller
      operationId: getBook
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/BookId"
      responses:
        "200":
          description: Book detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookDetail"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [books]
      summary: Edit metadata (creator or admin)
      operationId: patchBook
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/BookId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchBook"
      responses:
        "200":
          description: Updated metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookMeta"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [books]
      summary: |
        Delete the metadata entry (creator or admin). Metadata and file
        deletion are distinct operations: deleting files never touches the
        metadata and deleting metadata only succeeds once every file has
        been deleted explicitly.
      operationId: deleteBook
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/BookId"
      responses:
        "204":
          description: Deleted
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Book still has files; delete them first
          content:
            application/json:
              schema:
                type: object
                required: [error, files]
                properties:
                  error:
                    type: string
                  files:
                    type: array
                    description: Ids of the files still attached to the book
                    items:
                      type: string
...
components:
  parameters:
    BookId:
      name: id
      in: path
      required: true
      description: Book (metadata) id
      schema:
        type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
...
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing or invalid token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Not allowed to perform this action
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource does not exist (or share expired)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: Resource already exists
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    AcquireResult:
      type: object
      description: |
        Result of one acquisition (`POST /api/books`). When the plugin
        source declared >1 volumes, `series` is the created series and
        `books` holds one entry per 卷; otherwise `series` is null and
        `books` has exactly one entry. Always non-empty.
      required: [series, books]
      properties:
        series:
          $ref: "#/components/schemas/SeriesBrief"
          nullable: true
        books:
          type: array
          items:
            $ref: "#/components/schemas/BookDetail"
    BookDetail:
      type: object
      required: [book, files, series]
      properties:
        book:
          $ref: "#/components/schemas/BookMeta"
        files:
          description: Files of this book visible to the caller (empty for a metadata-only entry)
          type: array
          items:
            $ref: "#/components/schemas/FileMeta"
        series:
          $ref: "#/components/schemas/SeriesBrief"
          nullable: true
    BookListEntry:
      type: object
      required: [book, files, series]
      properties:
        book:
          $ref: "#/components/schemas/BookMeta"
        files:
          description: Files of this book visible to the caller
          type: array
          items:
            $ref: "#/components/schemas/FileMeta"
        series:
          description: Series this book belongs to (null = standalone)
          $ref: "#/components/schemas/SeriesBrief"
          nullable: true
...